+++ packages/apps/marmalade-marketplace/src/pages/tokens/index.tsx
@@ -1,87 +1,19 @@
-import { Heading, TextField, Button, Tabs, TabItem, Stack } from '@kadena/react-ui';
-import * as styles from "@/styles/create-token.css"
+import React, { useState, FormEvent } from 'react';
+import { Heading } from '@kadena/react-ui';
+import { Providers } from "@/providers/Providers";
 import Layout from '@/components/Layout';
+import CreateToken from '@/components/CreateToken';
 
-export default function CreateToken() {
+export default function TokenPages() {
+
   return (
-    <Stack flex={1} flexDirection="column">
-      <h1>Create Token</h1>
-      <div className={styles.twoColumnRow}>
-        <div className={styles.formSection}>
-          {}
-          <Button> Upload Image </Button>
-        </div>
-        <div className={styles.formSection}>
-          <form className={styles.verticalForm}>
-            <TextField label = "URI"/>
-            <TextField label = "Precision"/>
-            <TextField label = "Creation Guard"/>
-            <TextField label = "Chain"/>
-          </form>
-        </div>
-      </div>
-      <div className={styles.oneColumnRow}>
-        <div className={styles.formSection}>
-        <form className={styles.verticalForm}>
-              {/* Adjusted row with two columns */}
-              <div className={styles.checkboxRow}>
-                {/* First Column */}
-                <div className={styles.firstColumn}>
-                  <div className={styles.checkboxContainer}>
-                    <input className={styles.checkboxInput} type="checkbox" id="nonFungible" name="nonFungible" />
-                    <label className={styles.checkboxLabel} htmlFor="nonFungible">Non Fungible Policy</label>
-                  </div>
-                  <div className={styles.checkboxContainer}>
-                    <input className={styles.checkboxInput} type="checkbox" id="collection" name="collection" />
-                    <label className={styles.checkboxLabel} htmlFor="collection">Collection Policy</label>
-                  </div>
-                </div>
-                {/* Second Column */}
-                <div className={styles.secondColumn}>
-                  <div className={styles.checkboxContainer}>
-                    <input className={styles.checkboxInput} type="checkbox" id="guard" name="guard" />
-                    <label className={styles.checkboxLabel} htmlFor="guard">Guard Policy</label>
-                  </div>
-                  <div className={styles.checkboxContainer}>
-                    <input className={styles.checkboxInput} type="checkbox" id="royalty" name="royalty" />
-                    <label className={styles.checkboxLabel} htmlFor="royalty">Royalty Policy</label>
-                  </div>
-                </div>
-              </div>
-            </form>
-
-        </div>
-
-    <div className={styles.oneColumnRow}>
-    <div className={styles.formSection}>
-      <Tabs>
-        <TabItem title="Guards">
-          <form className={styles.verticalForm} title="Guards">
-            <TextField label = "URI Guard"/>
-            <TextField label = "Mint Guard"/>
-            <TextField label = "Burn Guard"/>
-            <TextField label = "Sale Guard"/>
-            <TextField label = "Transfer Guard"/>
-          </form>
-        </TabItem>
-        <TabItem title="Collection">
-          <form className={styles.verticalForm}>
-            <TextField label = "Collection Operator Guard"/>
-          </form>
-        </TabItem>
-        <TabItem title="Royalty">
-        <form className={styles.verticalForm}>
-            <TextField label = "Royalty Fungible"/>
-            <TextField label = "Royalty Creator"/>
-            <TextField label = "Royalty Guard"/>
-            <TextField label = "Royalty Rate"/>
-            <TextField label = "Royalty Fungible"/>
-          </form>
-        </TabItem>
-      </Tabs>
-      </div>
-      </div>
-      </div>
-    </Stack>
-  );
+    <Providers>
+      <Layout>
+        <Heading>
+            <title>Create Token</title>
+        </Heading>
+         <CreateToken/>
+      </Layout>
+    </Providers>
+);
 }