+++ packages/apps/tools/src/components/Global/CustomTree/CustomTree.css.ts
@@ -4,7 +4,6 @@ import { style } from '@vanilla-extract/css';
 export const containerStyle = style({
   display: 'flex',
   flexDirection: 'column',
-  height: '100%',
 });
 
 export const itemTitleStyle = style([