+++ packages/libs/react-ui/src/components/Layout/Grid/Grid.css.ts
@@ -1,19 +1,15 @@
 /* eslint @typescript-eslint/naming-convention: 0 */
-import { style, styleVariants } from '@vanilla-extract/css';
 import mapValues from 'lodash.mapvalues';
+import { style, styleVariants } from '../../../styles';
 import { breakpoints, mapToProperty } from '../../../styles/themeUtils';
 
-export const gridContainerClass = style([
-  {
-    display: 'grid',
-  },
-]);
+export const gridContainerClass = style({
+  display: 'grid',
+});
 
-export const gridItemClass = style([
-  {
-    gridColumnStart: 'auto',
-  },
-]);
+export const gridItemClass = style({
+  gridColumnStart: 'auto',
+});
 
 const columnCount: Record<number, number> = {
   1: 1,