+++ packages/libs/react-ui/src/components/Pagination/Pagination.css.ts
@@ -1,23 +1,18 @@
-import { style } from '@vanilla-extract/css';
-import { atoms, token } from '../../styles';
+import { style, token } from '../../styles';
 
-export const listClass = style([
-  atoms({
-    display: 'flex',
-    alignItems: 'center',
-    gap: 'xs',
-    padding: 'no',
-  }),
-  { listStyleType: 'none' },
-]);
+export const listClass = style({
+  display: 'flex',
+  alignItems: 'center',
+  gap: token('spacing.xs'),
+  padding: 0,
+  listStyleType: 'none',
+});
 
-export const pageNumButtonClass = style([
-  {
-    paddingBlock: `${token('spacing.xs')} !important`,
-    selectors: {
-      '&[data-current]': {
-        outline: `2px auto ${token('color.accent.brand.primary')} !important`,
-      },
+export const pageNumButtonClass = style({
+  paddingBlock: `${token('spacing.xs')} !important`,
+  selectors: {
+    '&[data-current]': {
+      outline: `2px auto ${token('color.accent.brand.primary')} !important`,
     },
   },
-]);
+});