+++ packages/apps/dev-wallet-desktop/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @kadena/dev-wallet-desktop
 
+## 0.0.11
+
+### Patch Changes
+
+- @kadena/dev-wallet\@0.1.3
+
 ## 0.0.10
 
 ### Patch Changes