+++ packages/apps/marmalade-marketplace/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @kadena/marmalade-marketplace
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies \[63bdbcb]
+- Updated dependencies \[83c3d58]
+- Updated dependencies \[15a492c]
+  - @kadena/react-ui\@0.12.0
+  - @kadena/client-utils\@0.8.0
+
 ## 0.0.2
 
 ### Patch Changes